requirements.txt
Flask==2.0.1
Pillow==9.0.0
pikepdf==5.0.0
PyMuPDF==1.19.6
python-docx==0.8.11
openpyxl==3.0.9
pandas==1.3.5
pdfplumber==0.6.3
reportlab==3.6.6